# core/clist: use insertion sort