--- a/core/lib/clist.c
+++ b/core/lib/clist.c
@@ -1,38 +1,9 @@
 /*
- * Copyright (C) 2017 Inria
- *               2017 Freie Universität Berlin
- *               2017 Kaspar Schleiser <kaspar@schleiser.de>
+ * Copyright (C) 2025 Marian Buschsieweke <marian.buschsieweke@posteo.net>
  *
  * This file is subject to the terms and conditions of the GNU Lesser
  * General Public License v2.1. See the file LICENSE in the top level
  * directory for more details.
- *
- * The code of _clist_sort() has been imported from
- * https://www.chiark.greenend.org.uk/~sgtatham/algorithms/listsort.html.
- * Original copyright notice:
- *
- * This file is copyright 2001 Simon Tatham.
- *
- * Permission is hereby granted, free of charge, to any person
- * obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without
- * restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or
- * sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following
- * conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT.  IN NO EVENT SHALL SIMON TATHAM BE LIABLE FOR
- * ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
- * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
  */
 
 /**
@@ -42,7 +13,7 @@
  * @file
  * @brief       clist helper implementations
  *
- * @author      Kaspar Schleiser <kaspar@schleiser.de>
+ * @author      Marian Buschsieweke <marian.buschsieweke@posteo.net>
  *
  * @}
  */
@@ -51,103 +22,60 @@
 
 clist_node_t *_clist_sort(clist_node_t *list, clist_cmp_func_t cmp)
 {
-    clist_node_t *p, *q, *e;
-    int insize, psize, qsize, i;
+    clist_node_t *sorted, *unsorted;
 
-    /*
-     * Silly special case: if `list' was passed in as NULL, return
-     * NULL immediately.
-     */
+    /* Special case: Empty list */
     if (!list) {
         return NULL;
     }
 
-    insize = 1;
-
-    while (1) {
-        clist_node_t *tail = NULL;
-        clist_node_t *oldhead = list;
-        p = list;
-        list = NULL;
-
-        int nmerges = 0;  /* count number of merges we do in this pass */
-
-        while (p) {
-            nmerges++;  /* there exists a merge to be done */
-            /* step `insize' places along from p */
-            q = p;
-            psize = 0;
-            for (i = 0; i < insize; i++) {
-                psize++;
-                q = (q->next == oldhead) ? NULL : q->next;
-                /* cppcheck-suppress nullPointer
-                 * (reason: possible bug in cppcheck 1.6x) */
-                if (!q) {
-                    break;
-                }
-            }
-
-            /* if q hasn't fallen off end, we have two lists to merge */
-            qsize = insize;
-
-            /* now we have two lists; merge them */
-            while (psize > 0 || (qsize > 0 && q)) {
-
-                /* decide whether next element of merge comes from p or q */
-                if (psize == 0) {
-                    /* p is empty; e must come from q. */
-                    e = q; q = q->next; qsize--;
-                    if (q == oldhead) {
-                        q = NULL;
-                    }
-                }
-                else if (qsize == 0 || !q) {
-                    /* q is empty; e must come from p. */
-                    e = p; p = p->next; psize--;
-                    if (p == oldhead) {
-                        p = NULL;
-                    }
-                }
-                else if (cmp(p, q) <= 0) {
-                    /* First element of p is lower (or same);
-                     * e must come from p. */
-                    e = p; p = p->next; psize--;
-                    if (p == oldhead) {
-                        p = NULL;
-                    }
-                }
-                else {
-                    /* First element of q is lower; e must come from q. */
-                    e = q; q = q->next; qsize--;
-                    if (q == oldhead) {
-                        q = NULL;
-                    }
-                }
-
-                /* add the next element to the merged list */
-                if (tail) {
-                    tail->next = e;
-                }
-                else {
-                    list = e;
-                }
-                tail = e;
-            }
-
-            /* now p has stepped `insize' places along, and q has too */
-            p = q;
-        }
-
-        /* cppcheck-suppress nullPointer
-         * (reason: tail cannot be NULL at this point, because list != NULL) */
-        tail->next = list;
-
-        /* If we have done only one merge, we're finished. */
-        if (nmerges <= 1) { /* allow for nmerges==0, the empty list case */
-            return tail;
-        }
-
-        /* Otherwise repeat, merging lists twice the size */
-        insize *= 2;
+    /* Special case: Only one item */
+    if (!list->next) {
+        return list;
     }
+
+    /* We break the circular list into two linear lists. The list sorted
+     * gets one element, all others go into unsorted. We then sort in one
+     * element from unsorted into sorted at a time using insertion sort.
+     * Note: Since sorted is one element long at start, it is naturally sorted,
+     *       as every one item list is inherently sorted.
+      */
+    sorted = list->next;
+    unsorted = sorted->next;
+    sorted->next = NULL;
+    list->next = NULL;
+
+sort_next_cycle:
+    while (unsorted) {
+        clist_node_t *iter = sorted;
+        clist_node_t **prev = &sorted;
+        while (iter) {
+            if (cmp(unsorted, iter) < 0) {
+                /* head of unsorted needs to be inserted before iter */
+                clist_node_t *tmp = unsorted->next;
+                *prev = unsorted;
+                unsorted->next = iter;
+                unsorted = tmp;
+                goto sort_next_cycle;
+            }
+            prev = &iter->next;
+            iter = iter->next;
+        }
+        /* head of unsorted needs to be inserted after iter */
+        clist_node_t *tmp = unsorted->next;
+        *prev = unsorted;
+        unsorted->next = NULL;
+        unsorted = tmp;
+    }
+
+    /* We now have in sorted a sorted linear list. Let's make that circular
+     * again and return the pointer to the last item */
+
+    clist_node_t *last = sorted;
+    while (last->next) {
+        last = last->next;
+    }
+    last->next = sorted;
+
+    return last;
 }