--- a/tests/bench/runtime_coreapis/main.c
+++ b/tests/bench/runtime_coreapis/main.c
@@ -19,6 +19,7 @@
  */
 
 #include <stdio.h>
+#include <string.h>
 
 #include "mutex.h"
 #include "benchmark.h"
@@ -26,7 +27,10 @@
 #include "thread_flags.h"
 
 #ifndef BENCH_RUNS
-#define BENCH_RUNS          (1000UL * 1000UL)
+#  define BENCH_RUNS          (1000UL * 1000UL)
+#endif
+#ifndef BENCH_CLIST_RUNS
+#  define BENCH_CLIST_RUNS (BENCH_RUNS / 1000UL)
 #endif
 
 static mutex_t _lock;
@@ -58,6 +62,71 @@ static void _flag_waitone(void)
     thread_flags_wait_one(_flag);
 }
 
+struct test_node {
+    clist_node_t list;
+    int value;
+};
+
+static clist_node_t clist;
+
+static void _build_test_clist(size_t len)
+{
+    static struct test_node nodes[128];
+
+    assert(len <= ARRAY_SIZE(nodes));
+
+    memset(nodes, 0, sizeof(nodes));
+    clist.next = NULL;
+
+    for (size_t i = 0; i < len; i++) {
+        clist_rpush(&clist, &nodes[i].list);
+    }
+}
+
+static void _insert_unsorted_data(clist_node_t *list)
+{
+    clist_node_t *last = list->next;
+    if (!last) {
+        return;
+    }
+    unsigned val = 128;
+
+    clist_node_t *cur = last->next;
+    while (cur != last) {
+        struct test_node *n = container_of(cur, struct test_node, list);
+
+        /* This is the worst case for most sorting algorithms: having the
+         * list perfectly sorted but in inverse direction. For real time, we
+         * only care about worst case and not about best case / average case */
+        n->value = val--;
+        cur = cur->next;
+    }
+}
+
+static int _cmp(clist_node_t *_lhs, clist_node_t *_rhs)
+{
+    struct test_node *lhs = container_of(_lhs, struct test_node, list);
+    struct test_node *rhs = container_of(_rhs, struct test_node, list);
+
+    return lhs->value - rhs->value;
+}
+
+static void _clist_sort_test(void)
+{
+    clist_node_t *list = &clist;
+
+    _insert_unsorted_data(list);
+    clist_sort(list, _cmp);
+}
+
+static void _bench_clist_sort(unsigned size)
+{
+    char name[32] = {};
+    snprintf(name, sizeof(name) - 1, "clist_sort() (%u nodes)", size);
+    _build_test_clist(size);
+    BENCHMARK_FUNC(name, BENCH_CLIST_RUNS, _clist_sort_test());
+}
+
 int main(void)
 {
     puts("Runtime of Selected Core API functions\n");
@@ -77,6 +146,10 @@ int main(void)
     puts("");
     BENCHMARK_FUNC("msg_try_receive()", BENCH_RUNS, msg_try_receive(&_msg));
     BENCHMARK_FUNC("msg_avail()", BENCH_RUNS, msg_avail());
+    puts("");
+    _bench_clist_sort(8);
+    _bench_clist_sort(32);
+    _bench_clist_sort(128);
 
     puts("\n[SUCCESS]");
     return 0;