### Contribution description

This replaces the previous clist sort implementation, which used a merge sort, with a simple insertion sort implementation.

The motivation is as follow:

1. The insertion sort is smaller (code size)
2. The performance benefit of the merge sort would only become noticeable on data sets that would be to large to comfortably handle on the MCUs RIOT targets
3. The previous implementation looked like a good candidate for the IOCCC, while the insertion sort is trivial to understand
4. Static analysis could provide for two statements an example control flow path that would trigger a `NULL` pointer dereferencation. The current code has no such defects.

### Testing procedure

1. There is a unit test that should still pass.
2. Adding `-fanalyzer` should now be happy with the code
3. The code should now be understandable without spending hours on a white board

### Issues/PRs references

None