@maribu 2025-04-11
I also improved the unit test a little bit to cover a few edge cases and to verify that the sorted list is actually sorted.