@maribu 2025-04-12
Depends on and includes:

- [ ] https://github.com/RIOT-OS/RIOT/pull/21402
- [ ] https://github.com/RIOT-OS/RIOT/pull/21403